requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclass-builder==1.2.0
feedparser==6.0.11
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
polyline==2.0.2
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
